added user input options, human readability option, csv creation

+++ script.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+echo "Create output folder of human readable data per file?"
+read -p "Enter y/n: " -n 1 -r $REPLY
+echo
+hr=false
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	hr=true
+fi
+
+
+
 #set up structures
 if [[ -d "./ascii" ]]
 then
@@ -10,9 +21,14 @@ then
 	rm -r ./output
 fi
 mkdir ./ascii
-mkdir ./output
 chmod 777 ascii
-chmod 777 output
+
+if [[ hr ]]
+then
+	mkdir ./output
+	chmod 777 output
+fi
+
 echo "*** Finished preprocessing"
 
 #Convert all pdfs to ascii text
@@ -25,7 +41,6 @@ do
 	pdftotext -layout pdfs/"$f.pdf" ./ascii/"$f.txt"
 done
 
-echo "In 4000 BC, I lost my 2019 manga collection. This was about 300 years ago, if I remember. Perhaps 65 AD." > ./ascii/check.txt
 echo "*** Finished Conversions"
 
 #Open docker container to hold stanford server on port 9000
@@ -33,7 +48,12 @@ echo "*** Finished Conversions"
 #docker run --rm -d -p 9000:9000 --name stanfordcorenlp anwala/stanfordcorenlp
 
 #create the csv for output
+if [[ -f "out.csv" ]]
+then
+	rm out.csv
+fi
 touch out.csv
+echo "period,dcterms:coverage_temporal,xsd:gDate_start,xsd:gDate_end" > out.csv
 
 #Parse for smithsonian trinomials, dates
 for FILE in ./ascii/*.txt
@@ -41,8 +61,12 @@ do
 	basename "$FILE"
 	#f="$(basename -- $FILE)"
 	f=$(basename "$FILE" .txt)
-	#python3 driver.py ascii/"$f" > ./output/"${f%.txt}.output"
-	python3 driver.py ascii/"$f.txt" 'y' > ./output/"$f.output"
+	if [[ hr ]]
+	then
+		python3 driver.py ascii/"$f.txt" "$REPLY" > ./output/"$f.output"
+	else
+		python3 driver.py ascii/"$f.txt" "$REPLY"
+	fi
 done
 echo "*** Finished parsing"
 
@@ -50,10 +74,13 @@ echo "*** Finished parsing"
 #docker rm -f stanfordcorenlp
 
 #Delete files with no useful data
-find ./output -size 0 -print -delete > ./output/deleted.txt
-find ./ascii -size 0 -print -delete > ./ascii/deleted.txt
-
 #Set permissions so all users can access (hack to fix root creation)
+if [[ -d "./output" ]]
+then
+	find ./output -size 0 -print -delete > ./output/deleted.txt
+	find ./output/ -type f -exec chmod 777 {} \;
+fi
+find ./ascii -size 0 -print -delete > ./ascii/deleted.txt
 find ./ascii/ -type f -exec chmod 777 {} \;
-find ./output/ -type f -exec chmod 777 {} \;
+
 echo "*** Finished setting permissions"
